Round weather values

+++ src/WeatherForecast/Presentation/Presentation.Api/Models/WeatherModel.cs
@@ -9,13 +9,13 @@ namespace Presentation.Api.Models
         public DateTime Date { get; set; }
         
         [DataMember(Name = "temp")]
-        public decimal Temperature { get; set; }
+        public int Temperature { get; set; }
         
         [DataMember(Name = "pressure")]
-        public decimal Pressure { get; set; }
+        public int Pressure { get; set; }
         
         [DataMember(Name = "humidity")]
-        public decimal Humidity { get; set; }
+        public int Humidity { get; set; }
         
         [DataMember(Name = "speed")]
         public decimal WindSpeed { get; set; }
